requirements.txt
django
djangorestframework
environs
mysqlclient
requests